Rename strategizer directory to planning
