+++ runtime/strategies/init-search.js
@@ -5,7 +5,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategy} from '../../strategizer/strategizer.js';
+import {Strategy} from '../../planning/strategizer.js';
 import {Recipe} from '../ts-build/recipe/recipe.js';
 import {assert} from '../../platform/assert-web.js';
