+++ runtime/ts/recipe/walker-base.ts
@@ -5,7 +5,7 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import {Strategizer} from '../../../strategizer/strategizer.js';
+import {Strategizer} from '../../../planning/strategizer.js';
 import {assert} from '../../../platform/assert-web.js';
 
 /**
